+++ internal/kibana/import_saved_objects/delete.go
@@ -1,0 +1,12 @@
+package import_saved_objects
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+)
+
+func (r *Resource) Delete(ctx context.Context, request resource.DeleteRequest, response *resource.DeleteResponse) {
+	tflog.Info(ctx, "Delete isn't supported for elasticstack_kibana_import_saved_objects")
+}