+++ internal/kibana/import_saved_objects/read.go
@@ -1,0 +1,12 @@
+package import_saved_objects
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+)
+
+func (r *Resource) Read(ctx context.Context, request resource.ReadRequest, response *resource.ReadResponse) {
+	tflog.Info(ctx, "Read isn't supported for elasticstack_kibana_import_saved_objects")
+}